refactor(web): split AttackBoard.tsx and add team color frames (ARC-427)
Each board now wears its team's color as a 2px border around the
PlayerSection plus a "● Team Name" pill next to the player name in
the header. This makes it obvious who's on which team when there are
3 or 4 teams in play, where a single "Teammate" badge isn't enough.

Also split the monolithic AttackBoard.tsx (~500 lines) into a folder:
  AttackBoard/
    index.tsx                 — wrapper component (109 lines)
    AttackPlayerBoard.tsx     — per-player board (301 lines)
    AttackBoardCell.tsx       — cell (85 lines)
    Pills.tsx                 — BadgePill + TeamPill (85 lines)
    cell-helpers.ts           — color/icon/anim helpers (30 lines)

Each file has one clear responsibility and stays well under the
500-line cap.
