+++ apps/web/src/widgets/SeaBattleGame/ui/Game.tsx
@@ -322,6 +322,7 @@ export const SeaBattleGame = memo(function SeaBattleGame({
             onAttack={attack}
             resolveDisplayName={resolveDisplayNameBound}
             teammateIds={teammateIds}
+            teams={teams}
           />
         )}
       </>
